+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # V 1.4.0
 Changes in this release:
+- Added meaningful error message when --venv points to a non-existing directory (#62)
 
 # V 1.3.0
 Changes in this release: