+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # V 1.4.0
 Changes in this release:
 - Added meaningful error message when --venv points to a non-existing directory (#62)
+- Ensure that cache is closed completely (#57)
 
 # V 1.3.0
 Changes in this release: