+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Changes in this release:
 - Added meaningful error message when --venv points to a non-existing directory (#62)
 - Ensure that cache is closed completely (#57)
+- Fix incompatibility with pytest 6.0.0
 
 # V 1.3.0
 Changes in this release: