+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Changes in this release:
 - Added meaningful error message when --venv points to a non-existing directory (#62)
 - Ensure that cache is closed completely (#57)
 - Fix incompatibility with pytest 6.0.0
-- Fixed plugin loading compatibility with compiler (#46, #49)
+- Fixed plugin loading compatibility with compiler's import mechanism (#46, #49)
+- Added `inmanta_plugins` fixture to make abstraction of required module reloading when the compiler project is reset (related to #49)
 
 # V 1.3.0
 Changes in this release: